requirements/common.txt
python-dotenv==1.0.0
django==4.1.7
djangorestframework==3.14
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
pyyaml==6.0
uritemplate==4.1.1
python-magic==0.4.27
psycopg2-binary==2.9.5
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-cleanup==7.0.0
redis==4.5.3
django-environ==0.10.0
whitenoise==6.4.0
django-colorfield==0.8.0
django-cors-headers==3.14.0
django-constance[redis]==2.9.1
channels[daphne]==4.0.0
azure-cognitiveservices-speech==1.25.0
channels_redis==4.0.0
yt-dlp==2023.3.4
openai==0.27.2
azure-identity==1.12.0
azure-mgmt-cognitiveservices==13.4.0
azure-mgmt-resource==23.0.0
azure-mgmt-monitor==6.0.0